Call to veneers in ARM mode
Since commit c6630ef92a21ca33af08b1faa03fba578e12315f, Call to a veneer
when the final symbol to be reached is thumb is made through a blx
instruction. This is a mistake since veneers are ARM instructions and
should thus be called with a simple bl. This commit prevent the bl ->
blx conversion when a veneer is used.