Stop setting -Wno-unused-result switch in Makefile
This commit revert commit 061b5799cc1feb15014da589ea98057deda14a23 and
subsequent commits to detect whether -Wno-unused-result is supported or
not by the compiler used to compile tcc. No warning about unused results
is issued in a normal build and thus this switch is only needed if
calling make with extra switches in CFLAGS or CPPFLAGS. It should thus
be added with the extra switches when calling make and not in the
Makefile.