Relicensing TinyCC
It has been discussed on the list whether it would be good
to relicense TinyCC under a more permissive BSD-like license.

The discussion started here:
http://lists.gnu.org/archive/html/tinycc-devel/2013-04/msg00052.html

Opinions varied but mostly were positive so it appears to
be worth to start the process and see how far we can get.

For that purpose I've committed a new file RELICENSING with the
suggested new license clause and a list for people to confirm
their agreement (or disagreement).

If you have contributed to TinyCC in the past, in particular if
you are one of the copyright owners for an entire file, please
add yourself to that file (rsp. replace the question mark) and
commit the change to the "mob" brancn with log message:

     Relicensing TinyCC

Thanks.